[UI] Embedded material css and fonts, more hooks refactoring
