+++ src/HealthChecks.UI/assets/index.html
@@ -4,9 +4,7 @@
 <head>
     <meta charset="utf-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1">
-    <title>Health Checks UI</title>
-    <link href="https://fonts.googleapis.com/icon?family=Material+Icons"
-      rel="stylesheet">
+    <title>Health Checks UI</title>    
     <link rel="stylesheet" href="#uiResourcePath#/healthchecksui-min.css" />
     #customstylesheets#
 </head>
