+++ src/HealthChecks.UI/client/index.tsx
@@ -1,3 +1,4 @@
+import "../assets/material.css";
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { App } from './App';
